# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Shop System API",
        "description": "API documentation for the Online Shop System. Supports public, customer, and admin endpoints.",
        "contact": {
            "email": "support@onlineshop.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders for authenticated user",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Paginated list of orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "operationId": "8114ac185cc1bf301d8862617e497098",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{order}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel a pending order",
                "operationId": "6b3a53e3f3f98b4e5da616fa85e4f37d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Only pending orders can be cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/products/{product}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get all reviews for a product",
                "operationId": "d1acae29ec1886ab7e982aa7021ba3d5",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product reviews with average rating"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit a review for a product",
                "operationId": "353de60b3ab807d62bb7a7a4f82d66a2",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "description": "Rating 1-5",
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review submitted"
                    },
                    "409": {
                        "description": "Already reviewed"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/{provider}/redirect": {
            "get": {
                "tags": [
                    "Social Auth"
                ],
                "summary": "Redirect to OAuth provider for authentication",
                "operationId": "7a69bff40b5bd8389eb17655fda6d94a",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "github"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to provider"
                    },
                    "400": {
                        "description": "Unsupported provider"
                    }
                }
            }
        },
        "/api/auth/{provider}/callback": {
            "get": {
                "tags": [
                    "Social Auth"
                ],
                "summary": "Handle OAuth provider callback",
                "operationId": "afd6512a8dab3944634f30b6fbf8b32d",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "google",
                                "github"
                            ]
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to frontend with token"
                    },
                    "400": {
                        "description": "Unsupported provider"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Bearer token obtained from login/register. Format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}